Revert `generation.py` rename and add `system_prompt` to `{Chat,Text}Generation`
